+++ src/main/java/insane96mcp/insanesurvivaloverhaul/module/combat/ISORunes.java
@@ -4,12 +4,17 @@ import insane96mcp.insanesurvivaloverhaul.InsaneSO;
 import insane96mcp.runeenchanting.runes.Rune;
 import insane96mcp.runeenchanting.setup.RERunes;
 import net.minecraft.data.tags.IntrinsicHolderTagsProvider;
+import net.minecraft.network.chat.Component;
+import net.minecraft.network.chat.MutableComponent;
 import net.minecraft.world.entity.EquipmentSlotGroup;
 import net.minecraft.world.entity.ai.attributes.AttributeModifier;
 import net.minecraft.world.item.Item;
 import net.neoforged.bus.api.IEventBus;
+import net.neoforged.neoforge.common.extensions.IAttributeExtension;
 import net.neoforged.neoforge.event.ItemAttributeModifierEvent;
 import net.neoforged.neoforge.registries.RegisterEvent;
+
+import javax.annotation.Nullable;
 
 /**
  * Registers the Critical and Armor Piercer runes into Rune Enchanting's rune registry.
@@ -40,6 +45,18 @@ public class ISORunes {
 		}
 
 		@Override
+		public @Nullable String getInfo() {
+			return "Critical chance: +%s%%, Critical damage: +%s%%";
+		}
+
+		@Override
+		public MutableComponent getInfoComponent() {
+			return Component.translatable(getInfoTranslationKey(),
+					IAttributeExtension.FORMAT.format(CriticalRework.criticalRuneBonusChance * CriticalRework.rune$enchantmentLevelEquivalent * 100),
+					IAttributeExtension.FORMAT.format(CriticalRework.criticalRuneBonusDamage * CriticalRework.rune$enchantmentLevelEquivalent * 100));
+		}
+
+		@Override
 		public void addItemsToApplicableTag(IntrinsicHolderTagsProvider.IntrinsicTagAppender<Item> appender) {}
 
 		@Override
@@ -60,6 +77,17 @@ public class ISORunes {
 		@Override
 		public String getDescription() {
 			return "Increases piercing damage";
+		}
+
+		@Override
+		public @Nullable String getInfo() {
+			return "Piercing damage: +%s per point of Attack Damage";
+		}
+
+		@Override
+		public MutableComponent getInfoComponent() {
+			return Component.translatable(getInfoTranslationKey(),
+					IAttributeExtension.FORMAT.format(PiercingDamage.armorPiercerDamagePerAttackDamage * PiercingDamage.rune$enchantmentLevelEquivalent));
 		}
 
 		@Override