+++ changelog.md
@@ -1,5 +1,6 @@
 # Upcoming
 * Fixed Max Health Death Penalty resetting on world join
+* Fixed Runes using wrong item tag
 
 # 5.8.0.0-beta
 * Tweaks