+++ changelog.md
@@ -5,6 +5,7 @@
 * Fixed Max Health Death Penalty resetting on world join
 * Fixed missing infos from runes
 * Fixed Runes using wrong item tag
+* Fixed regenerating absorption rendering on the wrong side
 
 # 5.8.0.0-beta
 * Tweaks